🚚 [e2e] move workshop solutions to examples
